refactor(application,persistence): move generic repository to dedicated folder and update namespaces
- Moved IGenericRepository interface to Interfaces/Repositories/Generic/ in the Application layer and updated its namespace.
- Moved GenericRepository implementation to Repositories/Generic/ in the Persistence layer and updated its namespace.
- Updated all repository interfaces and implementations to use the new namespace for IGenericRepository and GenericRepository.
- Updated DependencyInjection and relevant files to use the new namespaces.
- Improved project structure for better organization and scalability.
