+++ dotnet-backend/AirlineBookingSystem.Application/Interfaces/Repositories/IBookingStatusRepository.cs
@@ -1,3 +1,4 @@
+using AirlineBookingSystem.Application.Interfaces.Repositories.Generic;
 using AirlineBookingSystem.Domain.Entities;
 using AirlineBookingSystem.Shared.Enums;
