+++ dotnet-backend/AirlineBookingSystem.Application/Interfaces/Repositories/IPersonRepository.cs
@@ -1,3 +1,4 @@
+using AirlineBookingSystem.Application.Interfaces.Repositories.Generic;
 using AirlineBookingSystem.Domain.Entities;
 
 namespace AirlineBookingSystem.Application.Interfaces.Repositories;
