+++ dotnet-backend/AirlineBookingSystem.Persistence/Repositories/BookingRepository.cs
@@ -1,6 +1,7 @@
 using AirlineBookingSystem.Application.Interfaces.Repositories;
 using AirlineBookingSystem.Domain.Entities;
 using AirlineBookingSystem.Persistence.DbContext;
+using AirlineBookingSystem.Persistence.Repositories.Generic;
 using AirlineBookingSystem.Shared.Enums;
 using Microsoft.EntityFrameworkCore;
