+++ dotnet-backend/AirlineBookingSystem.Persistence/Repositories/CountryRepository.cs
@@ -1,6 +1,7 @@
 using AirlineBookingSystem.Application.Interfaces.Repositories;
 using AirlineBookingSystem.Domain.Entities;
 using AirlineBookingSystem.Persistence.DbContext;
+using AirlineBookingSystem.Persistence.Repositories.Generic;
 
 namespace AirlineBookingSystem.Persistence.Repositories;
